typo fix

+++ base/basetypes.h
@@ -7,7 +7,7 @@
 
 namespace base {
 
-typedef int32_t i16;
+typedef int16_t i16;
 typedef int32_t i32;
 typedef int64_t i64;
